+++ src/pages/designList/components/CategoryBar/CategoryBar.css.ts
@@ -3,9 +3,8 @@ import { style } from '@vanilla-extract/css';
 import { flexGenerator } from '@styles/generator.css';
 
 export const categoryBarContainer = style([
-  flexGenerator(),
+  flexGenerator('row', 'space-between'),
   {
     width: '100%',
-    gap: '1.6rem',
   },
 ]);
